requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.7.0
Brotli==1.1.0
browser-cookie3==0.20.1
bs4==0.0.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.6
cssselect==1.3.0
fake-useragent==2.2.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
imageio-ffmpeg==0.6.0
kaitaistruct==0.10
lxml==6.0.0
lz4==4.4.4
multidict==6.6.3
numpy==2.3.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pandas-stubs==2.3.0.250703
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
playwright==1.54.0
propcache==0.3.2
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydivert==2.1.0
pyee==13.0.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdfium2==4.30.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
requests==2.32.4
selenium==4.33.0
selenium-wire==5.1.0
setuptools==80.9.0
shadowcopy==0.0.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
WMI==1.5.1
wsproto==1.2.0
yarl==1.20.1
yt-dlp==2025.7.21
zstandard==0.23.0
